+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased](https://github.com/schemathesis/schemathesis/compare/v4.6.5...HEAD) - TBD
 
+### :bug: Fixed
+
+- False positive in `negative_data_rejection` check when multiple mutations conflict. [#3367](https://github.com/schemathesis/schemathesis/issues/3367)
+
 ## [4.6.5](https://github.com/schemathesis/schemathesis/compare/v4.6.4...v4.6.5) - 2025-12-03
 
 ### :bug: Fixed